refactor : statusText 함수 추가
